# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-bolierplate",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/fakultas/{kode_program_studi}": {
			"get": {
				"operationId": "GetDataProdi",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"message": "get-fuckultas"
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "kode_program_studi",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/fakultas/matakuliah-praktik/{kode_program_studi}": {
			"get": {
				"operationId": "GetMataKuliahPraktik",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"data": [
												{
													"sks": 4,
													"masa": "20202",
													"jumlah_pemeriksa": 2,
													"jumlah_nilai": 2,
													"persen_na1": 50,
													"persen_na2": 50,
													"status_praktek_mandiri": "0",
													"kode_jenis_formulasi": "NT",
													"kode_program_studi": "69",
													"kode_mtk": "MKPR4869",
													"nama_matakuliah": "Mata Kuliah PR",
													"folder_pedoman": "public/file-upload/(Form_Permintaan_dan_Penghapusan_Email)2.pdf",
													"file_pedoman": "(Form_Permintaan_dan_Penghapusan_Email)2.pdf"
												},
												{
													"sks": 4,
													"masa": "20202",
													"jumlah_pemeriksa": 2,
													"jumlah_nilai": 2,
													"persen_na1": 50,
													"persen_na2": 50,
													"status_praktek_mandiri": "0",
													"kode_jenis_formulasi": "NT",
													"kode_program_studi": "69",
													"kode_mtk": "MKPR4869",
													"nama_matakuliah": "Mata Kuliah PR",
													"folder_pedoman": "public/file-upload/(Form_Permintaan_dan_Penghapusan_Email)2.pdf",
													"file_pedoman": "(Form_Permintaan_dan_Penghapusan_Email)2.pdf"
												}
											],
											"success": true
										}
									}
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "kode_program_studi",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/fakultas/upload-pedoman": {
			"post": {
				"description": "Endpoint untuk upload pedoman pemeriksaan matakuliah",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "json",
									"example": {
										"data": {
											"sks": 4,
											"masa": "20202",
											"jumlah_periksa": 1,
											"persen_na1": 50,
											"persen_na2": 50,
											"status_praktek_mandiri": true,
											"kode_jenis_formula": "12",
											"kode_program_studi": "279",
											"kode_mtk": "PANG422",
											"nama_matakuliah": "Prak. Bio Kimia",
											"folder_pedoman": "folder/nama",
											"file_pedoman": "nama_file.pdf"
										}
									}
								}
							}
						}
					}
				},
				"consumes": [
					"multipart/form-data"
				],
				"parameters": [
					{
						"in": "formData",
						"name": "kode_matakuliah",
						"required": true,
						"type": "string"
					},
					{
						"in": "formData",
						"name": "pedoman",
						"required": true,
						"type": "file"
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}